requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Vector database and embeddings
chromadb==0.4.22
openai==1.10.0

# Git operations
gitpython==3.1.41

# Code parsing
tree-sitter==0.20.4
tree-sitter-c-sharp==0.20.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
